requirements/base.txt
# Installed from binary packages:
bcrypt
Pillow
psycopg2-binary
regex

# Python requirements:
bleach==2.1.3
Celery==4.4.7
channels==2.1.2
channels_redis==2.2.1
daphne==2.2.0
django-choices==1.6.0
django-cors-headers==2.2.0
django-filter==2.3.0
django-phonenumber-field==2.0.0
django-redis==4.10.0
django-stdimage==3.0.1
# DON'T BUMP django-storage-swift VERSION WITHOUT PROPER TESTING!
# It seems that in newer DSS releases auth token renewals might be broken.
# See: https://github.com/dennisv/django-storage-swift/issues/101#issuecomment-377138377
django-storage-swift==1.2.17
django-storages==1.6.6
django-waffle==0.14.0
Django==2.2.15
djangorestframework-gis==0.15.0
djangorestframework==3.12.1
drf-nested-routers==0.90.2
drf-extensions==0.4.0
html5lib==1.0.1
https://github.com/glemmaPaul/django-stdimage-serializer/archive/7a2ecc431a0afd4f292f79399660dd6cb2667018.zip
https://github.com/badygin-yakov/django-wysiwyg-redactor/archive/master.zip
idna==2.6
Markdown==2.6.11
python-slugify==1.2.5
redis==3.2.0
requests==2.18.4
django-debug-toolbar==2.2
certifi==2020.06.20
structlog
pyexcelerate==0.9.0
django-colorfield==0.3.2
nplusone==1.0.0
django-cachalot==2.3.3
django-nested-admin==3.3.3
pylightxl==1.55
